+++ lib/loadAST/basilVisitor.ml
@@ -9,7 +9,7 @@
 (** ASL visitor class *)
 
 open BasilIR.AbsBasilIR
-open Visitor
+open Common.Visitor
 
 class type basilVisitor = object
   method vdecl : declaration -> declaration visitAction
@@ -19,6 +19,8 @@ class type basilVisitor = object
   method vstmt : statement -> statement visitAction
   method vjump : jump -> jump visitAction
   method vtype : typeT -> typeT visitAction
+  method vexpr : expr -> expr visitAction
+  method vlvar : lVar -> lVar visitAction
 end
 
 let singletonVisitAction (a : 'a visitAction) : 'a list visitAction =
@@ -92,10 +94,95 @@ class virtual basilTreeVisitor (vis : #basilVisitor) =
       doVisit vis (vis#vblock b) next b
 
     method visit_statement (s : statement) : statement =
-      doVisit vis (vis#vstmt s) nochildren s
+      let next _ b =
+        match b with
+        | Assign (Assignment1 (o, expr)) ->
+            Assign (Assignment1 (self#visit_lvar o, self#visit_expr expr))
+        | SimulAssign alist ->
+            SimulAssign
+              (List.map
+                 (function
+                   | Assignment1 (o, expr) ->
+                       Assignment1 (self#visit_lvar o, self#visit_expr expr))
+                 alist)
+        | SLoad (lVar, endian, memory, addr, size) ->
+            let nlv = self#visit_lvar lVar in
+            let nadr = self#visit_expr addr in
+            if nlv <> lVar || nadr <> addr then
+              SLoad (nlv, endian, memory, nadr, size)
+            else b
+        | SStore (endian, bIdent, addr, value, size) ->
+            let nadr = self#visit_expr addr in
+            let nvalue = self#visit_expr value in
+            if nadr <> addr || nvalue <> value then
+              SStore (endian, bIdent, nadr, nvalue, size)
+            else b
+        | DirectCall (callLVars, bIdent, actual_params) ->
+            let params = mapNoCopy self#visit_expr actual_params in
+            if params <> actual_params then
+              DirectCall (callLVars, bIdent, params)
+            else b
+        | IndirectCall expr ->
+            let ne = self#visit_expr expr in
+            if ne <> expr then IndirectCall ne else b
+        | Assume (expr, attr) ->
+            let ne = self#visit_expr expr in
+            if ne <> expr then Assume (ne, attr) else b
+        | Guard (expr, attr) ->
+            let ne = self#visit_expr expr in
+            if ne <> expr then Guard (ne, attr) else b
+        | Assert (expr, attr) ->
+            let ne = self#visit_expr expr in
+            if ne <> expr then Assert (ne, attr) else b
+      in
+      doVisit vis (vis#vstmt s) next s
 
     method visit_jump (j : jump) : jump =
-      doVisit vis (vis#vjump j) nochildren j
+      let next _ j =
+        match j with
+        | Return params ->
+            let np = mapNoCopy self#visit_expr params in
+            if np <> params then Return np else j
+        | j -> j
+      in
+      doVisit vis (vis#vjump j) next j
+
+    method visit_expr (e : expr) =
+      let next _ e =
+        match e with
+        | GRVar rvar -> e
+        | LRVar rvar -> e
+        | BinaryExpr (binOp, l, r) ->
+            let nl = self#visit_expr l in
+            let nr = self#visit_expr r in
+            if nl <> l || nr <> r then BinaryExpr (binOp, nl, nr) else e
+        | UnaryExpr (unOp, l) ->
+            let nl = self#visit_expr l in
+            if nl <> l then UnaryExpr (unOp, l) else e
+        | ZeroExtend (intVal, expr) ->
+            let nl = self#visit_expr expr in
+            if nl <> expr then ZeroExtend (intVal, expr) else e
+        | SignExtend (intVal, expr) ->
+            let nl = self#visit_expr expr in
+            if nl <> expr then SignExtend (intVal, expr) else e
+        | Extract (upper, lower, expr) ->
+            let nl = self#visit_expr expr in
+            if nl <> expr then Extract (upper, lower, expr) else e
+        | Concat (l, r) ->
+            let nl = self#visit_expr l in
+            let nr = self#visit_expr r in
+            if nl <> l || nr <> r then Concat (nl, nr) else e
+        | Literal l -> Literal l
+        | Forall l -> Forall l
+        | Exists l -> Exists l
+        | OldExpr e -> OldExpr (self#visit_expr e)
+        | FunctionOp (i, param) ->
+            let param = mapNoCopy self#visit_expr param in
+            FunctionOp (i, param)
+      in
+      doVisit vis (vis#vexpr e) next e
+
+    method visit_lvar (e : lVar) = doVisit vis (vis#vlvar e) nochildren e
 
     method visit_type (x : typeT) : typeT =
       doVisit vis (vis#vtype x) nochildren x
@@ -110,6 +197,8 @@ class nopBasilVisitor : basilVisitor =
     method vstmt (_ : statement) = DoChildren
     method vjump (_ : jump) = DoChildren
     method vtype (_ : typeT) = DoChildren
+    method vexpr (_ : expr) = DoChildren
+    method vlvar (_ : lVar) = DoChildren
   end
 
 class forwardBasilvisitor (vis : #basilVisitor) =