+++ CHANGES.md
@@ -1,4 +1,15 @@
 
+## 0.1.0 (2025-07-1)
+
+features: 
+
+- semantic highlighting
+- code lens to view cfg using external dot process and viewer
+
+internally:
+
+- more complete AST and visitor
+
 ## 0.0.2 (2025-06-18)
 
 - update to basil il with sigils